E2E: check for uninstall pod init container successfully completed
...instead of pod phase "running" as we've observed sporadic failures
where the pod phase is still "pending" even though the init container
successfully completed. Also the operator log reported that the
associated daemonset was deleted. So uninstall worked correctly. The
pod status is cached in the E2E via an informer so it's likely the
phase update didn't occur b/c the pod was deleted or it hadn't
occurred yet.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>